refactor(mobile): share the source-scan helper between the token guards
Sonar's new-code duplication came in at 3.3% against a 3% gate. The two guard
tests each carried their own copy of the same recursive src walk and
offender loop.

findSourceFilesMatching(patterns, allowed) now lives in __tests__/setup and
both call it. Re-mutation-checked after the refactor: reintroducing a white
fill and a decorative-pink border still fails exactly one guard each.
