+++ apps/mobileAppYC/__tests__/setup/sourceScan.ts
@@ -1,0 +1,42 @@
+import {readdirSync, readFileSync, statSync} from 'fs';
+import {join} from 'path';
+
+/**
+ * Shared helper for the token guard tests, which each walk src/ looking for a
+ * banned pattern. Extracted so the two of them stop duplicating the walk.
+ */
+export const SRC_DIR = join(__dirname, '..', '..', 'src');
+
+const walk = (dir: string, out: string[] = []): string[] => {
+  for (const entry of readdirSync(dir)) {
+    const full = join(dir, entry);
+    if (statSync(full).isDirectory()) {
+      walk(full, out);
+    } else if (/\.tsx?$/.test(entry)) {
+      out.push(full);
+    }
+  }
+  return out;
+};
+
+/**
+ * Every src file whose contents match one of `patterns`, as paths relative to
+ * src/, skipping anything in `allowed`.
+ */
+export const findSourceFilesMatching = (
+  patterns: RegExp[],
+  allowed: ReadonlySet<string> = new Set(),
+): string[] => {
+  const offenders: string[] = [];
+  for (const file of walk(SRC_DIR)) {
+    const rel = file.slice(SRC_DIR.length + 1);
+    if (allowed.has(rel)) {
+      continue;
+    }
+    const body = readFileSync(file, 'utf8');
+    if (patterns.some(pattern => pattern.test(body))) {
+      offenders.push(rel);
+    }
+  }
+  return offenders;
+};
